add support for custom css files
If the user has a custom-styles folder in this plugin's path,
then any .css files in there will be listed as selectable HTML
themes, with the name "Custom - <filename>", and will be inserted
into the email's footer to style it.

The body was also made into standard HTML5 with a DOCTYPE, header,
body, and footer.